+++ app/tidningen/[year]/[issue]/[page]/error.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { GenericCatchBoundary } from '@/components/CatchBoundary';
+
+export default function ErrorPage({ error }: { error: Error }) {
+  return <GenericCatchBoundary caught={error} />;
+}
